requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
